+++ client/horizontal-menu/change-background-image.js
