+++ client/horizontal-menu/template.hbs
@@ -1,5 +1,5 @@
 {{#if elements}}
     <ol class="outer-container horizontal-list btn-group no-left-padding">
-        {{> ./_listItem elements=elements}}
+        {{> ./list-item elements=elements}}
     </ol>
 {{/if}}
