+++ client/marker-list-modal/template.hbs
@@ -30,7 +30,7 @@
                             </div>
                         {{/if}} --}}
                         <div class="map-marker-preview-container">
-                            {{> ./map-marker-modal-preview}}
+                            {{> ../preview-modal/template}}
                         </div>
                     </div>
                 {{/each}}
