+++ client/marker-table/template.hbs
