+++ client/preview-modal/index.unit.test.js
@@ -2,16 +2,16 @@ const testHelpers = require('@quoin/node-test-helpers');
 const expect = testHelpers.expect;
 const proxyquire = testHelpers.proxyquire.noPreserveCache();
 
-describe('client/lib/utilities/map-marker-modal-preview', () => {
+describe('client/utilities/map-marker-modal-preview', () => {
     let stubs;
     let mapMarkerModalPreview;
 
     beforeEach(() => {
         stubs = {
-            './../templates/map-marker-modal-preview.hbs': testHelpers.stub()
+            './../preview-modal/template.hbs': testHelpers.stub()
         };
 
-        mapMarkerModalPreview = proxyquire.noCallThru().load(require.resolve('./map-marker-modal-preview'), stubs);
+        mapMarkerModalPreview = proxyquire.noCallThru().load(require.resolve('./index'), stubs);
     });
 
     it('should be a function', () => {
