+++ client/vertical-menu/template.hbs
@@ -1,4 +1,4 @@
-{{> ./_selectRowLink selectableLinks=selectableLinks}}
+{{> types-dropdown selectableLinks=selectableLinks}}
 {{#if elements}}
     <div class="outer-container vertical-list btn-group-vertical no-left-padding full-width" role="group" aria-label="">
         {{#each elements}}
