+++ client/vertical-menu/types-dropdown.hbs
