+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@warp-works/warpjs-map-plugin",
-  "version": "0.2.1",
+  "version": "0.2.2",
   "description": "WarpJS map browser plugin",
   "main": "index.js",
   "scripts": {
@@ -42,14 +42,14 @@
     "@warp-works/warpjs-utils": "latest",
     "babel-core": "6.26.0",
     "babel-loader": "7.1.2",
-    "babel-preset-env": "1.6.0",
+    "babel-preset-env": "1.6.1",
     "babel-preset-stage-0": "6.24.1",
-    "eslint": "4.8.0",
+    "eslint": "4.10.0",
     "eslint-plugin-html": "3.2.2",
-    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-import": "2.8.0",
     "eslint-plugin-json": "1.2.0",
-    "eslint-plugin-node": "5.2.0",
-    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-node": "5.2.1",
+    "eslint-plugin-promise": "3.6.0",
     "eslint-plugin-standard": "3.0.1",
     "express": "4.16.2",
     "grunt": "1.0.1",
@@ -57,7 +57,7 @@
     "grunt-contrib-less": "1.4.1",
     "grunt-eslint": "20.1.0",
     "grunt-webpack": "3.0.2",
-    "handlebars": "4.0.10",
+    "handlebars": "4.0.11",
     "handlebars-loader": "1.6.0",
     "jquery": "3.2.1",
     "jsdom": "11.3.0",
